Moved galaxy_counts tests to separate test script.
